Save failed dicom files from users in dedicated group
If a dicom file in an upload session from a user in the dedicated DICOM_DATA_CREATORS_GROUP_NAME group cannot be processed these files will not be removed at the end of the file processing step. The original uploaded files can then be downloaded via the django-admin.

+++ app/grandchallenge/cases/tasks.py
@@ -9,6 +9,8 @@ from typing import Sequence, Tuple
 from uuid import UUID
 
 from celery import shared_task
+from django.conf import settings
+from django.contrib.auth.models import Group
 from django.db import transaction
 
 from grandchallenge.algorithms.models import Job
@@ -485,10 +487,20 @@ def _handle_unconsumed_files(
 
 
 def _delete_session_files(*, session_files):
+    dicom_group = Group.objects.get(
+        name=settings.DICOM_DATA_CREATORS_GROUP_NAME
+    )
+    users = dicom_group.user_set.values_list("username", flat=True)
     for file in session_files:
         try:
             if file.staged_file_id:
                 saf = StagedAjaxFile(file.staged_file_id)
+                if (
+                    not file.consumed
+                    and Path(file.filename).suffix == ".dcm"
+                    and file.creator.username in users
+                ):
+                    continue
                 file.staged_file_id = None
                 saf.delete()
             file.save()
